requirements.txt
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0
timm>=1.0.0
torchmetrics>=1.7.0
pandas>=2.2.0
numpy>=2.2.0
scikit-learn>=1.6.0
scipy>=1.15.0
Pillow>=11.2.0
tqdm>=4.67.0
matplotlib>=3.10.0
seaborn>=0.13.0
PyYAML>=6.0
requests>=2.32.0
lightning-utilities>=0.11.0 # Опционально, если используется
# huggingface-hub # Опционально, если используется
# safetensors # Опционально, если используется